move form logic to another Component
